Add tcpfo argument to tcp_connect to setsockopt for resumption
We want to send the TCPFO cookie only when we resume.
I think we have to set the FASTOPEN_CONNECT flag on the (new) socket
before actually connect()ing.

tcp_connect does both, create a new socket and connect() to the target.
We have to be in between those two calls and this we squeeze ourselves
into that function with a conditional.

Currently, we want to perform full handshakes without any cookie
whatsoever, so we can update all callers to not set the flag while the
resuming callers set it.